requirements.txt
streamlit
google-generativeai
python-dotenv
reportlab
langchain
langchain-community
langchain-openai
faiss-cpu
